Replace string literals with typed Action variants in widget audit linkage
Add SetObjectPosition, SetObjectSize and SetObjectRotation to the Action
enum as Document-kind placeholders. Wire them through kind(), name(),
identifier(), requires_selection(), prepare_command and all exhaustive
tests. Update make_numeric_field in properties.rs to accept an Action and
derive the linkage string via identifier(), eliminating fragile raw-string
action names.

To keep gauss-core within the Whitaker module-max-lines limit, extract the
impl Action block into a new action_impl module.
